Refactor HitsInternal classes. HitsInternalNoLock.
